requirements.txt
plotly==5.8.0
pymongo==4.1.1
websocket==0.2.1
backtrader==1.9.76.123
certifi==2022.5.18.1
chardet==3.0.4
click==8.1.3
cssselect==1.1.0
cycler==0.11.0
idna==3.3
Jinja2==3.1.2
Keras==2.9.0
kur==0.7.0
lxml==4.9.0
MarkupSafe==2.0.1
matplotlib==3.5.2
nose==1.3.7
numpy==1.22.4
crypto==1.4.1
pycryptodome==3.14.1
pydub==0.25.1
pyparsing==3.0.9
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.26
python-speech-features==0.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.8.1
six==1.16.0
Theano==1.0.5
tornado==6.1
tqdm==4.64.0
urllib3==1.26.9